+++ src/Policies/BasePolicy.php
@@ -4,7 +4,7 @@ namespace TCG\Voyager\Policies;
 
 use Illuminate\Auth\Access\HandlesAuthorization;
 use TCG\Voyager\Facades\Voyager;
-use TCG\Voyager\Models\User;
+use TCG\Voyager\Contracts\User;
 
 class BasePolicy
 {
@@ -13,7 +13,7 @@ class BasePolicy
     /**
      * Determine if the given user can browse the model.
      *
-     * @param \TCG\Voyager\Models\User $user
+     * @param \TCG\Voyager\Contracts\User $user
      * @param  $model
      *
      * @return bool
@@ -23,13 +23,13 @@ class BasePolicy
         $dataType = Voyager::model('DataType');
         $dataType = $dataType->where('model_name', get_class($model))->first();
 
-        return Voyager::can('browse_'.$dataType->name);
+        return $user->hasPermission('browse_'.$dataType->name);
     }
 
     /**
      * Determine if the given model can be viewed by the user.
      *
-     * @param \TCG\Voyager\Models\User $user
+     * @param \TCG\Voyager\Contracts\User $user
      * @param  $model
      *
      * @return bool
@@ -39,13 +39,13 @@ class BasePolicy
         $dataType = Voyager::model('DataType');
         $dataType = $dataType->where('model_name', get_class($model))->first();
 
-        return Voyager::can('read_'.$dataType->name);
+        return $user->hasPermission('read_'.$dataType->name);
     }
 
     /**
      * Determine if the given model can be edited by the user.
      *
-     * @param \TCG\Voyager\Models\User $user
+     * @param \TCG\Voyager\Contracts\User $user
      * @param  $model
      *
      * @return bool
@@ -55,13 +55,13 @@ class BasePolicy
         $dataType = Voyager::model('DataType');
         $dataType = $dataType->where('model_name', get_class($model))->first();
 
-        return Voyager::can('edit_'.$dataType->name);
+        return $user->hasPermission('edit_'.$dataType->name);
     }
 
     /**
      * Determine if the given user can create the model.
      *
-     * @param \TCG\Voyager\Models\User $user
+     * @param \TCG\Voyager\Contracts\User $user
      * @param  $model
      *
      * @return bool
@@ -71,13 +71,13 @@ class BasePolicy
         $dataType = Voyager::model('DataType');
         $dataType = $dataType->where('model_name', get_class($model))->first();
 
-        return Voyager::can('add_'.$dataType->name);
+        return $user->hasPermission('add_'.$dataType->name);
     }
 
     /**
      * Determine if the given model can be deleted by the user.
      *
-     * @param \TCG\Voyager\Models\User $user
+     * @param \TCG\Voyager\Contracts\User $user
      * @param  $model
      *
      * @return bool
@@ -87,6 +87,6 @@ class BasePolicy
         $dataType = Voyager::model('DataType');
         $dataType = $dataType->where('model_name', get_class($model))->first();
 
-        return Voyager::can('delete_'.$dataType->name);
+        return $user->hasPermission('delete_'.$dataType->name);
     }
 }
