+++ src/Policies/PostPolicy.php
@@ -3,14 +3,14 @@
 namespace TCG\Voyager\Policies;
 
 use TCG\Voyager\Facades\Voyager;
-use TCG\Voyager\Models\User;
+use TCG\Voyager\Contracts\User;
 
 class PostPolicy extends BasePolicy
 {
     /**
      * Determine if the given model can be viewed by the user.
      *
-     * @param \TCG\Voyager\Models\User $user
+     * @param \TCG\Voyager\Contracts\User $user
      * @param  $model
      *
      * @return bool
@@ -23,7 +23,7 @@ class PostPolicy extends BasePolicy
         // Does this post belong to the current user?
         $current = $user->id === $model->author_id ? true : false;
 
-        $permission = Voyager::can('read_'.$dataType->name);
+        $permission = $user->hasPermission('read_'.$dataType->name);
 
         return $current || $permission;
     }
@@ -31,7 +31,7 @@ class PostPolicy extends BasePolicy
     /**
      * Determine if the given model can be edited by the user.
      *
-     * @param \TCG\Voyager\Models\User $user
+     * @param \TCG\Voyager\Contracts\User $user
      * @param  $model
      *
      * @return bool
@@ -44,7 +44,7 @@ class PostPolicy extends BasePolicy
         // Does this post belong to the current user?
         $current = $user->id === $model->author_id ? true : false;
 
-        $permission = Voyager::can('edit_'.$dataType->name);
+        $permission = $user->hasPermission('edit_'.$dataType->name);
 
         return $current || $permission;
     }
@@ -52,7 +52,7 @@ class PostPolicy extends BasePolicy
     /**
      * Determine if the given model can be deleted by the user.
      *
-     * @param \TCG\Voyager\Models\User $user
+     * @param \TCG\Voyager\Contracts\User $user
      * @param  $model
      *
      * @return bool
@@ -65,7 +65,7 @@ class PostPolicy extends BasePolicy
         // Does this post belong to the current user?
         $current = $user->id === $model->author_id ? true : false;
 
-        $permission = Voyager::can('delete_'.$dataType->name);
+        $permission = $user->hasPermission('delete_'.$dataType->name);
 
         return $current || $permission;
     }
