+++ src/Policies/SettingPolicy.php
@@ -3,20 +3,20 @@
 namespace TCG\Voyager\Policies;
 
 use TCG\Voyager\Facades\Voyager;
-use TCG\Voyager\Models\User;
+use TCG\Voyager\Contracts\User;
 
 class SettingPolicy extends BasePolicy
 {
     /**
      * Determine if the given user can browse the model.
      *
-     * @param \TCG\Voyager\Models\User $user
+     * @param \TCG\Voyager\Contracts\User $user
      * @param  $model
      *
      * @return bool
      */
     public function browse(User $user, $model)
     {
-        return Voyager::can('browse_settings');
+        return $user->hasPermission('browse_settings');
     }
 }
